Update Azure OIDC troubleshooting guide to use correct response type
This is a follow on to https://github.com/pivotal-cf/docs-identity/pull/157
where it was found that the prior recommendation to use `id_token`
response type was found to be incorrect.

The `id_token` response type causes UAA to negotiate an implicit grant
flow, which was broken in recent releases of UAA. Using the `code`
response type causes UAA and Azure to perform the Authorization Code grant flow,
which is more secure and functions as expected.

[#173017981](https://www.pivotaltracker.com/story/show/173017981)

Signed-off-by: Margo Crawford <margaretc@vmware.com>